// src/lib/api/models/ClientDetails.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AppointmentService } from './AppointmentService';
export type ClientDetails = {
  id?: number;
  service?: AppointmentService;
  firstname?: string;
  lastname?: string;
  email?: string;
  phoneNumber?: string;
  address?: string;
  comment?: string;
  createdAt?: string;
};